Move mock execution class to conftest
